+++ analyzers/dataframe/src/JetFlavourUtils.cc
@@ -49,7 +49,7 @@ namespace FCCAnalyses {
         throw std::runtime_error("Invalid index requested for jet flavour weight.");
       rv::RVec<float> out;
       for (const auto& jet_weights : jets_weights) {
-        if (weight >= jet_weights.size())
+        if (weight >= static_cast<int>(jet_weights.size()))
           throw std::runtime_error("Flavour weight index exceeds the number of weights registered.");
         out.emplace_back(jet_weights.at(weight));
       }
